Extract AddressBookModel
